lang/funcs: add "all" function
# What

This commit adds an `all` function to Terraform configuration. A
reason we might want this function is because it will enable more
powerful custom variable validations. For example:

```hcl
variable "amis" {
  type = list(object({
    id = string
  }))

  validation {
    condition = (all([
      for a in var.amis : length(a.id) > 4 && substr(a.id, 0, 4) == "ami-"
    ]))
    error_message = "The ID of at least one AMI was invalid."
  }
}

If this is not the intended use of custom variable validation, the `all`
function may still prove useful in other contexts where we, for example,
create a resource conditionally.
```